Move group functions to their own files
