+++ man/vec_group.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/dictionary.R
+% Please edit documentation in R/group.R
 \name{vec_group}
 \alias{vec_group}
 \alias{vec_group_id}
