+++ src/init.c
@@ -24,10 +24,10 @@ extern SEXP vec_split_id(SEXP);
 extern SEXP vec_split(SEXP, SEXP);
 extern SEXP vctrs_unique_loc(SEXP);
 extern SEXP vctrs_count(SEXP);
-extern SEXP vctrs_group_id(SEXP);
-extern SEXP vctrs_group_rle(SEXP);
 extern SEXP vctrs_id(SEXP);
 extern SEXP vctrs_n_distinct(SEXP);
+extern SEXP vctrs_group_id(SEXP);
+extern SEXP vctrs_group_rle(SEXP);
 extern SEXP vctrs_equal(SEXP, SEXP, SEXP);
 extern SEXP vctrs_duplicate_all(SEXP);
 extern SEXP vctrs_equal_na(SEXP);
@@ -113,10 +113,10 @@ static const R_CallMethodDef CallEntries[] = {
   {"vctrs_split",                      (DL_FUNC) &vec_split, 2},
   {"vctrs_duplicated_any",             (DL_FUNC) &vctrs_duplicated_any, 1},
   {"vctrs_count",                      (DL_FUNC) &vctrs_count, 1},
-  {"vctrs_group_id",                   (DL_FUNC) &vctrs_group_id, 1},
-  {"vctrs_group_rle",                  (DL_FUNC) &vctrs_group_rle, 1},
   {"vctrs_id",                         (DL_FUNC) &vctrs_id, 1},
   {"vctrs_n_distinct",                 (DL_FUNC) &vctrs_n_distinct, 1},
+  {"vctrs_group_id",                   (DL_FUNC) &vctrs_group_id, 1},
+  {"vctrs_group_rle",                  (DL_FUNC) &vctrs_group_rle, 1},
   {"vctrs_size",                       (DL_FUNC) &vctrs_size, 1},
   {"vctrs_dim",                        (DL_FUNC) &vec_dim, 1},
   {"vctrs_dim_n",                      (DL_FUNC) &vctrs_dim_n, 1},
